Simplified Navbar Structure
Simplified navbar structure for accessability and maintainability. Replaced custom MenuLink with built-in NavLink, and implemented active style using the aria-current attribute created by NavLink. Restyled the navbar to look like tabs, not great but serviceable for now. The Download CSV button should be moved, once it's moved the tab bar can be floated to the right (as it is, the download button would move the entire tab bar back and forth).